requirements.txt
fastapi==0.110.2
python-decouple==3.8
uvicorn==0.29.0
